app/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-adminlte2==0.4.3
django-bootstrap-v5==1.0.5
django-braces==1.14.0
django-extra-views==0.8.0
django-filter==1.0.0
django-mathfilters==1.0.0
django-simple-history==3.0.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2
djangorestframework==3.5.3
future==0.18.2
google-api-core==2.0.1
google-api-python-client==2.20.0
google-auth==2.0.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
importlib-metadata==3.10.1
oauth2client==1.5.2
protobuf==3.17.3
psycopg2>=2.8.6,<2.9